+++ src/components/InstancesList/InstancesList.js
@@ -166,6 +166,7 @@ class InstancesList extends React.Component {
     if (this.getSelectedBrowseOption() || this.getExecutedBrowseQuery()) {
       this.setState({
         optionSelected: this.getQIndexFromParams(),
+        isSingleResult: false,
       });
     } else {
       this.setState({
@@ -967,10 +968,10 @@ class InstancesList extends React.Component {
     const isHandleOnNeedMore = Object.values(browseModeOptions).includes(optionSelected) ? handleOnNeedMore : null;
 
     const onChangeIndex = (e) => {
-      const { value } = e.target;
-      this.setState({ optionSelected: value });
-      const isSingle = value !== browseModeOptions.CALL_NUMBERS || value !== browseModeOptions.SUBJECTS;
-      this.setState({ isSingleResult: isSingle });
+      this.setState({ optionSelected: e.target.value });
+      if (e.target.value === browseModeOptions.CALL_NUMBERS || e.target.value === browseModeOptions.SUBJECTS) {
+        this.setState({ isSingleResult: false });
+      } else this.setState({ isSingleResult: true });
     };
 
     const browseFilter = () => {